why was doc_len a uint8_t?

+++ main.c
@@ -92,7 +92,7 @@ int main (void) {
     size_t cursor = 0;
     uint8_t *doc;
     doc = malloc(sizeof(uint8_t)*1920);
-    uint8_t doc_len = 0;
+    size_t doc_len = 0;
     
     // Main Editor Loop
     printf("Entering main loop...\n");
@@ -109,7 +109,7 @@ int main (void) {
             doc[doc_len] = key;
             doc_len++;
             printf("Doc = %s\n", doc);
-            printf("Len = %zu\n",doc_len);
+            printf("Len = %zu\n", doc_len);
             if (doc_len % 5 == 0) {
                 set_screen_damage();
             }
